few improvements on the test
