+++ bigbluebutton-html5/imports/ui/components/actions-bar/actions-dropdown/component.jsx
@@ -277,7 +277,7 @@ class ActionsDropdown extends PureComponent {
           : intl.formatMessage(intlMessages.activateTimerStopwatchLabel),
         key: this.timerId,
         onClick: () => this.handleTimerClick(),
-        dataTest: 'timer',
+        dataTest: 'timerStopWatchFeature',
       });
     }
