+++ bigbluebutton-tests/playwright/core/elements.js
@@ -161,10 +161,10 @@ exports.virtualBackgrounds = 'div[data-test="virtualBackground"]';
 exports.learningDashboard = 'li[data-test="learningDashboard"]';
 
 // Timer
-exports.timer = 'li[data-test="timer"]';
+exports.timerStopwatchFeature = 'li[data-test="timerStopWatchFeature"]';
 exports.timerCurrent = 'span[data-test="timerCurrent"]';
 exports.startStopTimer = 'button[data-test="startStopTimer"]';
-exports.resetTimer = 'button[data-test="reset"]';
+exports.resetTimerStopwatch = 'button[data-test="resetTimerStopWatch"]';
 exports.timerButton = 'button[data-test="timer"]';
 exports.timerIndicator = 'div[data-test="timeIndicator"]';
 exports.stopwatch = 'button[data-test="stopwatch"]';
