+++ bigbluebutton-tests/playwright/parameters/util.js
@@ -102,13 +102,6 @@ async function checkShortcutsArray(test, shortcut) {
   }
 }
 
-async function timeInSeconds(locator) {
-  const texto = await locator.innerText();
-  const [horas, minutos, segundos] = texto.split(':').map(Number);
-    const tempoEmSegundos = horas * 3600 + minutos * 60 + segundos;
-    return tempoEmSegundos;
-}
-
 exports.zoomIn = zoomIn;
 exports.zoomOut = zoomOut;
 exports.poll = poll;
@@ -121,4 +114,3 @@ exports.encodeCustomParams = encodeCustomParams;
 exports.getAllShortcutParams = getAllShortcutParams;
 exports.checkAccesskey = checkAccesskey;
 exports.checkShortcutsArray = checkShortcutsArray;
-exports.timeInSeconds = timeInSeconds;
