+++ bigbluebutton-tests/playwright/user/user.spec.js
@@ -4,6 +4,7 @@ const { MultiUsers } = require('./multiusers');
 const { GuestPolicy } = require('./guestPolicy');
 const { LockViewers } = require('./lockViewers');
 const { MobileDevices } = require('./mobileDevices');
+const { Timer } = require('./timer');
 const motoG4 = devices['Moto G4'];
 const iPhone11 = devices['iPhone 11'];
 
@@ -298,4 +299,10 @@ test.describe.parallel('User', () => {
       await mobileDevices.chatPanelNotAppearOnMobile();
     });
   });
+});
+
+test('Timer', async ({ browser, context, page })=> {
+  const timer = new Timer(browser, context);
+  await timer.initModPage(page, true);
+  await timer.timerTest();
 });
